[Modules] Use PCMCache to manage memory buffers for pcm files.
We create a PCMCache class to manage memory buffers associated with pcm files.
With implicit modules, we currently use lock files to make sure we are making
progress when a single clang invocation builds a module file and then loads the
module file immediately after. This implementation requires lock file for
correctness and causes all kinds of locking issues.

This patch tries to use PCMCache to share the content between writes and reads
within a single clang invocation, so we can remove the correctness-side of the
lock file, and only use lock file for performance i.e to reduce chance of
building the same pcm file from multiple clang invocations.

We also add ThreadContext in PCMCache to diagnose cases where an ancestor
validates a module and later on, a child thread tries to invalidate it. Without
the patch, this scenario will cause use-after-free of the FileEntry held by the
ancestor.

https://reviews.llvm.org/D28299

rdar://problem/19720334