[APINotes] Always provide an unversioned SwiftName of some kind. (#81)
If the active API notes set a Swift name for a declaration, but no
other versions do, it's assumed that that's the name in every version,
because the AST is indistinguishable from just applying the attribute
in the header. Fix this by explicitly adding an inactive, unversioned
removal attribute for SwiftName in otherwise ambiguous cases.

(This really relates to /all/ annotations in the "SwiftVersions"
section, and particularly those that are usually not written in
headers, but it's particularly bad for SwiftName because the Swift
compiler actually looks at the inactive versions for that one. At
some point it may be worth trying to generalize this, though.)

rdar://problem/31826517