[CFG] [analyzer] Avoid modeling C++17 constructors that aren't fully supported.
Not enough work has been done so far to ensure correctness of construction
contexts in the CFG when C++17 copy elision is in effect, so for now we
should drop construction contexts in the CFG and in the analyzer when
they seem different from what we support anyway.

This includes initializations with conditional operators and return values
across multiple stack frames.

Differential Revision: https://reviews.llvm.org/D44854

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@328893 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit a83ee8e3d18648165fcdc7782981578669fa1c63)