[upstream-with-swift] Increase DIAG_SIZE_FRONTEND to 101
The swift-clang/upstream-with-swift branch currently has 1 extra diag
compared to upstream Clang. It looks like upstream is now right at that
limit of 100, so the swift-clang version is failing. I'm expecting the
upstream version to bump up the limit soon (i.e., the next time someone
tries to add a frontend diagnostic) but it would be good to keep this
version higher by 1.