[OPENMP][BLOCKS]Fix PR38923: reference to a global variable is captured by a block.
Added checks for capturing of the variable in the block when trying to
emit correct address for the variable with the reference type. This
extra check allows correctly identify the variables that are not
captured in the block context.

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@340181 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 5b14cddec6bdb8da8f375dfe0c0910ff970a0e05)