[APINotes] API-noted attributes should not be implicit (#189)
Even though they don't have valid source locations, they /are/ a
developer-specified part of the AST, and so it makes sense to include
them in the printed representation. This also brings back the tests
that were disabled in 8ab16db.

Note that the SwiftVersionedAttr and SwiftVersionedRemovalAttr
attributes are still considered implicit, since they really are part
of the compiler infrastructure and not something the developer chose
to put in themselves. (They also don't have a printed representation
anyway.)

rdar://problem/40296113